Small change to logic for ComponentRequirement.requiresAPassedInstance.
Previously, it would decide that a module instance must be passed if the module had any abstract method that was not a binding method. But this isn't quite true: if the module's binding methods are all abstract and/or static, then the abstract non-binding method doesn't matter at all; instance state does not affect the module.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=236879817